refactor(key-management): split repair missing keys dialog
